requirements.txt
cleanvision==0.3.5
huggingface_hub==0.20.2
lightning==2.1.2
numpy==1.23.5
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.2.0
rich==13.7.0
streamlit==1.30.0
timm==0.9.2
torch==2.1.0
torchmetrics==0.10.3
torchvision==0.16.0
tqdm==4.65.0
transformers==4.34.1
